+++ tests/index.test.ts
@@ -1,16 +1,39 @@
-const nock = require('nock')
-const core = require('@actions/core')
-const { Toolkit } = require('actions-toolkit')
+import nock from 'nock'
+import * as core from '@actions/core'
+import { Toolkit } from 'actions-toolkit'
+import { Signale } from 'signale'
+import { createAnIssue } from '../src/action'
+
+function generateToolkit() {
+  const tools = new Toolkit({
+    logger: new Signale({ disabled: true })
+  })
+
+  jest.spyOn(tools.log, 'info')
+  jest.spyOn(tools.log, 'error')
+  jest.spyOn(tools.log, 'success')
+
+  // Turn core.setOutput into a mocked noop
+  jest.spyOn(core, 'setOutput')
+    .mockImplementation(() => {})
+
+  // Turn core.setFailed into a mocked noop
+  jest.spyOn(core, 'setFailed')
+    .mockImplementation(() => {})
+
+  tools.exit.success = jest.fn() as any
+  tools.exit.failure = jest.fn() as any
+
+  return tools
+}
 
 describe('create-an-issue', () => {
-  let actionFn, tools, params
+  let tools: Toolkit
+  let params: any
 
   beforeEach(() => {
-    Toolkit.run = jest.fn(fn => { actionFn = fn })
-    require('../src')
-
     nock('https://api.github.com')
-      .post(/\/repos\/.*\/.*\/issues/).reply(200, (_, body) => {
+      .post(/\/repos\/.*\/.*\/issues/).reply(200, (_, body: any) => {
         params = body
         return {
           title: body.title,
@@ -19,26 +42,7 @@ describe('create-an-issue', () => {
         }
       })
 
-    tools = new Toolkit({
-      logger: {
-        info: jest.fn(),
-        success: jest.fn(),
-        warn: jest.fn(),
-        debug: jest.fn(),
-        error: jest.fn()
-      }
-    })
-
-    // Turn core.setOutput into a mocked noop
-    jest.spyOn(core, 'setOutput')
-      .mockImplementation(() => {})
-
-    // Turn core.setFailed into a mocked noop
-    jest.spyOn(core, 'setFailed')
-      .mockImplementation(() => {})
-
-    tools.exit.success = jest.fn()
-    tools.exit.failure = jest.fn()
+    tools = generateToolkit()
 
     // Ensure that the filename input isn't set at the start of a test
     delete process.env.INPUT_FILENAME
@@ -48,11 +52,10 @@ describe('create-an-issue', () => {
   })
 
   it('creates a new issue', async () => {
-    tools.log.success = jest.fn()
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
 
     // Verify that the outputs were set
     expect(core.setOutput).toHaveBeenCalledTimes(2)
@@ -63,55 +66,55 @@ describe('create-an-issue', () => {
   it('creates a new issue from a different template', async () => {
     process.env.INPUT_FILENAME = '.github/different-template.md'
     tools.context.payload = { repository: { owner: { login: 'JasonEtco' }, name: 'waddup' } }
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with some template variables', async () => {
     process.env.INPUT_FILENAME = '.github/variables.md'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with assignees, labels and a milestone', async () => {
     process.env.INPUT_FILENAME = '.github/kitchen-sink.md'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with assignees and labels as comma-delimited strings', async () => {
     process.env.INPUT_FILENAME = '.github/split-strings.md'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with an assignee passed by input', async () => {
     process.env.INPUT_ASSIGNEES = 'octocat'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with multiple assignees passed by input', async () => {
     process.env.INPUT_ASSIGNEES = 'octocat, JasonEtco'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.success).toHaveBeenCalled()
-    expect(tools.log.success.mock.calls).toMatchSnapshot()
+    expect((tools.log.success as any).mock.calls).toMatchSnapshot()
   })
 
   it('creates a new issue with a milestone passed by input', async () => {
     process.env.INPUT_MILESTONE = '1'
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(params.milestone).toBe('1')
     expect(tools.log.success).toHaveBeenCalled()
@@ -123,7 +126,7 @@ describe('create-an-issue', () => {
       .get(/\/search\/issues.*/).reply(200, {
         items: []
       })
-      .post(/\/repos\/.*\/.*\/issues/).reply(200, (_, body) => {
+      .post(/\/repos\/.*\/.*\/issues/).reply(200, (_, body: any) => {
         params = body
         return {
           title: body.title,
@@ -134,7 +137,7 @@ describe('create-an-issue', () => {
 
     process.env.INPUT_UPDATE_EXISTING = 'true'
 
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.log.info).toHaveBeenCalledWith('No existing issue found to update')
     expect(tools.log.success).toHaveBeenCalled()
@@ -149,7 +152,7 @@ describe('create-an-issue', () => {
       .patch(/\/repos\/.*\/.*\/issues\/.*/).reply(200, {})
     process.env.INPUT_UPDATE_EXISTING = 'true'
 
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(params).toMatchSnapshot()
     expect(tools.exit.success).toHaveBeenCalled()
   })
@@ -164,7 +167,7 @@ describe('create-an-issue', () => {
         message: 'Updating issue failed'
       })
 
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(tools.exit.failure).toHaveBeenCalled()
   })
 
@@ -175,9 +178,9 @@ describe('create-an-issue', () => {
         message: 'Validation error'
       })
 
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(tools.log.error).toHaveBeenCalled()
-    expect(tools.log.error.mock.calls).toMatchSnapshot()
+    expect((tools.log.error as any).mock.calls).toMatchSnapshot()
     expect(tools.exit.failure).toHaveBeenCalled()
   })
 
@@ -189,9 +192,9 @@ describe('create-an-issue', () => {
         errors: [{ foo: true }]
       })
 
-    await actionFn(tools)
+    await createAnIssue(tools)
     expect(tools.log.error).toHaveBeenCalled()
-    expect(tools.log.error.mock.calls).toMatchSnapshot()
+    expect((tools.log.error as any).mock.calls).toMatchSnapshot()
     expect(tools.exit.failure).toHaveBeenCalled()
   })
 })
